crud warranty, store

+++ src/main/java/com/web/prj/services/user/UserServiceImpl.java
@@ -36,17 +36,19 @@ public class UserServiceImpl implements UserService {
         if (existingUser.isPresent()) {
             throw new AppException(ErrorCode.USER_EXISTED);
         }
-
         User user = userMapper.toEntity(input);
-
-        Long lastUserId = userRepository.findMaxId().orElse(0L);
-        String accountId = "TK" + String.format("%06d", lastUserId + 1);
-        user.setAccountId(accountId);
-
         user.setRole(
                 roleRepository.findByRoleId(ROLE.USER.getLevel()).orElseThrow(() -> new AppException(ErrorCode.ROLE_NOT_FOUND))
         );
 
+        String phone = input.getPhone();
+        if(userRepository.existsByPhone(phone)) {
+            throw new AppException(ErrorCode.RECORD_EXISTED);
+        }
+
+        Long lastUserId = userRepository.findMaxId().orElse(0L);
+        String accountId = "TK" + String.format("%06d", lastUserId + 1);
+        user.setAccountId(accountId);
 
         return userMapper.toResponse(userRepository.save(user));
     }
@@ -56,23 +58,30 @@ public class UserServiceImpl implements UserService {
     public UserResponse updateUser(UserRequest input) {
         User user = userRepository.findByEmail(input.getEmail())
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+        Role role = roleRepository.findById(input.getRoleId())
+                .orElseThrow(() -> new AppException(ErrorCode.ROLE_NOT_FOUND));
 
-        user.setName(input.getName());
         if (input.getAvatar() != null && !input.getAvatar().isEmpty()) {
             user.setAvatar(input.getAvatar());
         }
+        user.setName(input.getName());
         user.setGender(input.getGender());
         user.setBirth(input.getBirth());
         user.setPhone(input.getPhone());
         user.setActive(input.isActive());
         user.setAddress(input.getAddress());
+        user.setRole(role);
+
         return userMapper.toResponse(userRepository.save(user));
     }
 
     @Override
     public void deleteUser(Long id) {
-        userRepository.findById(id)
+        User user = userRepository.findById(id)
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
+        if(user.getOrders() != null && !user.getOrders().isEmpty()) {
+            throw new AppException(ErrorCode.RECORD_UNDELETED);
+        }
         userRepository.delete(id);
     }
 
@@ -84,24 +93,6 @@ public class UserServiceImpl implements UserService {
         return userMapper.toResponse(user);
     }
 
-    @Override
-    public UserResponse grantRole(Long roleId, Long userId) {
-        Role role = roleRepository.findById(roleId)
-                .orElseThrow(() -> new AppException(ErrorCode.ROLE_NOT_FOUND));
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
-
-        user.setRole(role);
-        return userMapper.toResponse(userRepository.save(user));
-    }
-
-    @Override
-    public UserResponse applyMember(Long memberId, Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
-        ;
-        return null;
-    }
 
     @Override
     public PageResponse<UserResponse> findAllByPageAndFilter(Pageable pageable, Optional<String> filter) {
